add all tasks

+++ src/redux/contactsOps.js
@@ -1,11 +1,24 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
 import axios from "axios";
 
-axios.defaults.baseURL = "https://6853ba4ca2a37a1d6f499c64.mockapi.io/";
+axios.defaults.baseURL = "https://connections-api.goit.global";
+
+const setAuthHeader = (token) => {
+  axios.defaults.headers.common.Authorization = `Bearer ${token}`;
+};
 
 export const fetchContacts = createAsyncThunk(
   "contacts/fetchAll",
   async (_, thunkAPI) => {
+    const state = thunkAPI.getState();
+    const token = state.auth.token;
+
+    if (!token) {
+      return thunkAPI.rejectWithValue("No auth token");
+    }
+
+    setAuthHeader(token);
+
     try {
       const { data } = await axios.get("/contacts");
       return data;
@@ -18,6 +31,15 @@ export const fetchContacts = createAsyncThunk(
 export const addContact = createAsyncThunk(
   "contacts/addContact",
   async (contact, thunkAPI) => {
+    const state = thunkAPI.getState();
+    const token = state.auth.token;
+
+    if (!token) {
+      return thunkAPI.rejectWithValue("No auth token");
+    }
+
+    setAuthHeader(token);
+
     try {
       const { data } = await axios.post("/contacts", contact);
       return data;
@@ -30,6 +52,15 @@ export const addContact = createAsyncThunk(
 export const deleteContact = createAsyncThunk(
   "contacts/deleteContact",
   async (id, thunkAPI) => {
+    const state = thunkAPI.getState();
+    const token = state.auth.token;
+
+    if (!token) {
+      return thunkAPI.rejectWithValue("No auth token");
+    }
+
+    setAuthHeader(token);
+
     try {
       await axios.delete(`/contacts/${id}`);
       return id;
